remove unnessesary casts

+++ Source/AI/JobDriver_RechargeEnergy.cs
@@ -57,7 +57,7 @@ namespace Androids
                 if (compPowerBattery != null)
                 {
                     compPowerBattery.DrawPower(PowerNetEnergyDrainedPerTick);
-                    energyNeed.CurLevel += energyNeed.MaxLevel / (float)MaxTicksSpentCharging;
+                    energyNeed.CurLevel += energyNeed.MaxLevel / MaxTicksSpentCharging;
                 }
                 ticksSpentCharging++;
             };
